+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to the "blockman" extension will be documented in this file.
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
-## [Unreleased]
+## [1.2.5] - 2021-08-22
 
-- Initial release
+-   Yaml language support
+-   Increased char number limit on each line from 500 chars to 3000 chars.
+-   At the Blockman installation event -> "editor.inlayHints.enabled": false,
+
+## [1.0.1] ... [1.2.0]
+
+-   Non documented changes
+
+## [1.0.0] - 2021-05-28
+
+-   Initial release